chore: major changes (#93)
