+++ App.js
@@ -3,15 +3,16 @@ import React from "react";
 import { StatusBar } from "expo-status-bar";
 import { StyleSheet } from "react-native";
 import Board from "./components/Board/Board.jsx";
-import Main from "./components/Main/Main.jsx";
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
-import Register from "./components/Register/Register.jsx";
-import Login from "./components/Login/Login.jsx";
-import Profile from "./components/Profile/Profile.jsx";
-import GameStat from "./components/Profile/GameStat.jsx";
-import OnlineGameView from "./views/gameView/onlineGameView/OnlineGameView.jsx";
-import ReviewGameView from "./views/gameView/gameReview/ReviewGameView.jsx";
+import OnlineGameScreen from "./screens/GameScreens/OnlineGameScreen/OnlineGameScreen";
+import Login from "./screens/LoginScreen/Login";
+import Register from "./screens/RegisterScreen/Register";
+import Profile from "./screens/ProfileScreen/Profile";
+import GameStat from "./screens/ProfileScreen/GameStat";
+import HomeScreen from "./screens/HomeScreen/HomeScreen";
+import ReviewGameScreen from "./screens/GameScreens/GameReviewScreen/ReviewGameScreen";
+
 const Stack = createStackNavigator();
 
 export default function App() {
@@ -23,19 +24,18 @@ export default function App() {
           headerShown: false
         }}
       >
-        <Stack.Screen name="Chess" component={Main} />
+        <Stack.Screen name="Chess" component={HomeScreen} />
         <Stack.Screen
           name="localGame"
           component={Board}
           initialParams={{ isLocalGame: true }}
         />
-        <Stack.Screen name="onlineGame" component={OnlineGameView} />
-        <Stack.Screen name="reviewGame" component={ReviewGameView} />
+        <Stack.Screen name="onlineGame" component={OnlineGameScreen} />
+        <Stack.Screen name="reviewGame" component={ReviewGameScreen} />
         <Stack.Screen name="login" component={Login} />
         <Stack.Screen name="register" component={Register} />
         <Stack.Screen name="profile" component={Profile} />
         <Stack.Screen name="stats" component={GameStat} />
-        {/* <Stack.Screen name="reviewGame" component={GameReview} /> */}
       </Stack.Navigator>
       <StatusBar style="auto" />
     </NavigationContainer>
