+++ socket/websocket.js
@@ -1,5 +1,5 @@
 import { BACKEND_WEB_SOCKET_URL as url } from "@env";
-console.log(url);
+
 const createSocket = async function (userId) {
   console.log(url);
   return new Promise(function (resolve, reject) {
